+++ gemma-core/src/main/java/ubic/gemma/persistence/service/common/quantitationtype/QuantitationTypeService.java
@@ -40,6 +40,14 @@ import java.util.Set;
 public interface QuantitationTypeService extends BaseService<QuantitationType>, FilteringVoEnabledService<QuantitationType, QuantitationTypeValueObject> {
 
     /**
+     * @see QuantitationTypeDao#getVectorTypes()
+     */
+    Collection<Class<? extends DataVector>> getVectorTypes();
+
+    @Nullable
+    QuantitationType loadById( Long value, ExpressionExperiment ee );
+
+    /**
      * Find a quantitation type by ID and vector type.
      * <p>
      * While the QT can be retrieved uniquely by ID, the purpose of this method is to ensure that it also belongs to a
@@ -62,6 +70,9 @@ public interface QuantitationTypeService extends BaseService<QuantitationType>, 
     @Nullable
     @Secured({ "IS_AUTHENTICATED_ANONYMOUSLY", "ACL_SECURABLE_READ" })
     QuantitationType find( ExpressionExperiment ee, QuantitationType quantitationType, Class<? extends DataVector> dataVectorTypes );
+
+    @Secured({ "IS_AUTHENTICATED_ANONYMOUSLY", "ACL_SECURABLE_READ" })
+    QuantitationType findByName( ExpressionExperiment ee, String name ) throws NonUniqueQuantitationTypeByNameException;
 
     /**
      * @see QuantitationTypeDao#findByNameAndVectorType(ExpressionExperiment, String, Class)