+++ gemma-rest/src/main/java/ubic/gemma/rest/util/args/DatasetArgService.java
@@ -9,6 +9,7 @@ import ubic.gemma.core.analysis.preprocess.OutlierDetails;
 import ubic.gemma.core.analysis.preprocess.OutlierDetectionService;
 import ubic.gemma.core.search.*;
 import ubic.gemma.model.common.description.AnnotationValueObject;
+import ubic.gemma.model.common.quantitationtype.QuantitationType;
 import ubic.gemma.model.common.quantitationtype.QuantitationTypeValueObject;
 import ubic.gemma.model.common.search.SearchSettings;
 import ubic.gemma.model.expression.arrayDesign.ArrayDesignValueObject;
@@ -177,7 +178,23 @@ public class DatasetArgService extends AbstractEntityArgService<ExpressionExperi
      */
     public List<BioAssayValueObject> getSamples( DatasetArg<?> arg ) {
         ExpressionExperiment ee = service.thawLite( this.getEntity( arg ) );
-        List<BioAssayValueObject> bioAssayValueObjects = baService.loadValueObjects( ee.getBioAssays(), true );
+        return createSampleVos( ee, ee.getBioAssays() );
+    }
+
+    /**
+     * Obtain a collection of BioAssays that represent the experiments samples for a particular quantitation type.
+     */
+    public List<BioAssayValueObject> getSamples( DatasetArg<?> datasetArg, QuantitationType qt ) {
+        ExpressionExperiment ee = getEntity( datasetArg );
+        BioAssayDimension bad = service.getBioAssayDimension( ee, qt );
+        if ( bad == null ) {
+            throw new NotFoundException( "There are no assays associated to " + qt + "." );
+        }
+        return createSampleVos( ee, bad.getBioAssays() );
+    }
+
+    private List<BioAssayValueObject> createSampleVos( ExpressionExperiment ee, Collection<BioAssay> samples ) {
+        List<BioAssayValueObject> bioAssayValueObjects = baService.loadValueObjects( samples, true );
         Collection<OutlierDetails> outliers = outlierDetectionService.getOutlierDetails( ee );
         if ( outliers != null ) {
             Set<Long> predictedOutlierBioAssayIds = outliers.stream()
@@ -261,5 +278,10 @@ public class DatasetArgService extends AbstractEntityArgService<ExpressionExperi
                     return new BioAssayValueObject( ba, null, sourceAssay, false, true );
                 } )
                 .collect( Collectors.toList() );
+    }
+
+    public QuantitationType getPreferredQuantitationType( DatasetArg<?> datasetArg ) {
+        return service.getPreferredQuantitationType( getEntity( datasetArg ) )
+                .orElseThrow( () -> new NotFoundException( "No preferred quantitation type found for dataset with ID " + datasetArg + "." ) );
     }
 }